lab2 done

+++ lab2/DebugExercise/DebugExercise1.java
@@ -6,8 +6,8 @@ package DebugExercise;
  */
 public class DebugExercise1 {
     public static int divideThenRound(int top, int bottom) {
-        int quotient = top / bottom;
-        int result = Math.round(quotient);
+        double quotient = (double) top / bottom;
+        int result = (int) Math.round(quotient);
         return result;
     }
 
@@ -26,5 +26,6 @@ public class DebugExercise1 {
         int b3 = 4;
         int result3 = divideThenRound(t3, b3);
         System.out.println("round(" + t3 + "/" + b3 + ")=" + result3);
+
     }
 }
